feat: Add a back button on trackDetails
