+++ package.json
@@ -11,8 +11,10 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "dotenv": "^16.0.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-icons": "^4.3.1",
     "react-infinite-scroller": "^1.2.5",
     "react-query": "^3.34.19",
     "react-router-dom": "^6.2.1",
