+++ src/components/card/styles.css
@@ -53,8 +53,15 @@
   margin: 0 10px 0 0;
 }
 
-.contentBx h2 {
+.artistNameContainer {
+  max-height: 60px;
+  overflow: hidden;
+}
+
+.contentBx h3 {
   margin-bottom: 10px;
+  max-height: 50px;
+  overflow: hidden;
 }
 
 .contentBx {
