+++ src/components/trackDetails/styles.css
@@ -70,3 +70,14 @@
   color: inherit;
   text-decoration: inherit;
 }
+
+.backButton {
+  position: absolute;
+  max-height: 20px;
+  margin-top: 10px;
+  width: 20px;
+  height: 20px;
+  left: 0;
+  padding-left: 10px;
+  cursor: pointer;
+}
